+++ sdk-node/src/errors/index.ts
@@ -5,6 +5,7 @@ export {
   ThreeCommonConflictError,
   ThreeCommonConnectionError,
   ThreeCommonNotFoundError,
+  ThreeCommonPaymentRequiredError,
   ThreeCommonPermissionError,
   ThreeCommonRateLimitError,
   ThreeCommonServerError,